Rename 3. Data Cleaning Project Queries.sql to 2. Data Cleaning Project Queries.sql
In this project, I improved the quality of the redundancy data of many companies from around the world for the period from 2020 to 2023 by performing important cleansing work. 

Using SQL scripts, I eliminated duplicates and standardised key attributes such as "company", "industry" and "country". 

I also took care of invalid entries and, in particular, reformatted the date formats. 

To optimise the data set, I removed irrelevant columns and thus improved the usability of the data for analyses. 

This significantly improved the integrity and efficiency of the discharge data.