Rename Data Exploration Queries.sql to 3. Data Exploration Queries.sql
In this project, I performed exploratory data analysis (EDA) on a layoff dataset to identify key trends and outliers. I ran SQL queries to determine the maximum number of layoffs and scored the layoffs as a percentage to identify the fully downsized workforce. 

Grouped by company, location and time, I gained insights from the aggregated redundancies and used ranking functions to analyse the year-on-year impact. 

A rolling total calculation of monthly layoffs also highlighted temporal fluctuations in job cuts and provided important insights for economic strategy and policy making.